SVN-Revision: 6185

+++ projects/core/src/main/db/migrate/011_Phase2_construction_1/036_modify_mandatory_field_defs.groovy
@@ -1,9 +1,28 @@
 class ModifyMandatoryFieldDefs extends edu.northwestern.bioinformatics.bering.Migration {
 	void up() {
-		renameColumn("mandatory_field_defs", "mandatory", "mandatory_temp");
+		// add new mandatory column(integer) and drop existing mandatory (boolean)
+
+		if (databaseMatches('postgres')) {
+			// rename existing boolean column to mandatory_temp
+			// not droping in postgres because we need the existing data to be migrated to new column
+			
+			renameColumn("mandatory_field_defs", "mandatory", "mandatory_temp");
+		} else {
+			// drop boolean column if it is not postgres, because no migration needed
+			
+			dropColumn("mandatory_field_defs", "mandatory");
+		}
+		
+		//add new column with integer		
+		
 		addColumn("mandatory_field_defs", "mandatory", "integer", nullable: false, defaultValue: 0);
-		execute("update mandatory_field_defs set mandatory = (case when (mandatory_temp = true) then 1 else 0 end)");
-		dropColumn("mandatory_field_defs", "mandatory_temp");
+		
+		if (databaseMatches('postgres')) {
+			// migrate old data and drop temp column
+			
+			execute("update mandatory_field_defs set mandatory = (case when (mandatory_temp = true) then 1 else 0 end)");
+			dropColumn("mandatory_field_defs", "mandatory_temp");
+		}
 	}
 	
 	void down() {
